api/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
openai
httpx
passlib
python-jose[cryptography]
redis
pydantic[email]>=2.0.0
python-multipart
bcrypt
groq

